requirements.txt
google-adk
google-cloud-bigquery
google-cloud-aiplatform
google-genai
requests
python-dotenv
PyYAML
fastapi
uvicorn
pydantic
pydantic-settings
psycopg2-binary
mcp
httpx
redis
